refactor difficulty overview
splits file into multiple components, adds a FilterableLineChart component that wraps mantine's LineCharts and adds proper filtering by clicking legends
